requirements.txt
-i https://pypi.org/simple/
boto3>=1.34.29
botocore>=1.34.29
click==7.1.2
docutils==0.15.2
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.11.3
jmespath==0.10.0
markupsafe==1.1.1
python-dateutil==2.8.1 ; python_version >= '2.7'
s3transfer>=0.10.0
six==1.15.0
urllib3==1.25.11 ; python_version >= '3.4'
werkzeug==1.0.1
zipstream-new>=1.1.8
python-dotenv>=1.0.0
s3fs>=0.4.2